Moved IndexPage into its own file
