+++ network-api/networkapi/wagtailpages/utils.py
@@ -1,8 +1,23 @@
+import re
+
 from itertools import chain
 from django.apps import apps
 from django.db.models import Count
 from django.utils.translation import gettext
 from sentry_sdk import capture_exception
+
+
+def titlecase(s):
+    '''
+    See https://docs.python.org/3.7/library/stdtypes.html#str.title
+    for why this definition exists (basically: apostrophes)
+    '''
+    return re.sub(
+        r"[A-Za-z]+('[A-Za-z]+)?",
+        lambda mo: mo.group(0)[0].upper() +
+        mo.group(0)[1:].lower(),
+        s
+    )
 
 
 def set_main_site_nav_information(page, context, homepage_class_name):
